**JavaScript DOM manipulation: selectors**
Selectors allow developers to reference HTML elements stored as nodes in the DOM with the purpose of later manipulation.

To select an element you can use multiple methods that use the next identifiers: classes, id and element type.  This methods are applied over the document object that holds the DOM or an already selected reference stored in a variable to access its children.